Make recursive structurer a separate file.
