+++ angr/analyses/decompiler/structuring/__init__.py
@@ -1,1 +1,2 @@
-from .structurer import RecursiveStructurer, Structurer
+from .structurer import Structurer
+from .recursive_structurer import RecursiveStructurer
